fixed fuchsia by moving impeller logic into an impeller compiled file
