+++ lib/ui/painting/image.cc
@@ -8,19 +8,13 @@
 #include <limits>
 
 #if IMPELLER_SUPPORTS_RENDERING
-#include "flutter/impeller/renderer/texture.h"
+#include "flutter/lib/ui/painting/image_encoding_impeller.h"
 #endif
 #include "flutter/lib/ui/painting/image_encoding.h"
 #include "third_party/tonic/converter/dart_converter.h"
 #include "third_party/tonic/dart_args.h"
 #include "third_party/tonic/dart_binding_macros.h"
 #include "third_party/tonic/dart_library_natives.h"
-
-// Must be kept in sync with painting.dart.
-enum ColorSpace {
-  kSRGB,
-  kExtendedSRGB,
-};
 
 namespace flutter {
 
@@ -49,15 +43,7 @@ int CanvasImage::colorSpace() {
     return ColorSpace::kSRGB;
   } else if (image_->impeller_texture()) {
 #if IMPELLER_SUPPORTS_RENDERING
-    const impeller::TextureDescriptor& desc =
-        image_->impeller_texture()->GetTextureDescriptor();
-    switch (desc.format) {
-      case impeller::PixelFormat::kB10G10R10XR:  // intentional_fallthrough
-      case impeller::PixelFormat::kR16G16B16A16Float:
-        return ColorSpace::kExtendedSRGB;
-      default:
-        return ColorSpace::kSRGB;
-    }
+    return ImageEncodingImpeller::GetColorSpace(image_->impeller_texture());
 #endif  // IMPELLER_SUPPORTS_RENDERING
   }
