+++ lib/ui/painting/image_encoding_impeller.cc
@@ -163,4 +163,16 @@ void ImageEncodingImpeller::ConvertImageToRaster(
   });
 }
 
+int ImageEncodingImpeller::GetColorSpace(
+    const std::shared_ptr<impeller::Texture>& texture) {
+  const impeller::TextureDescriptor& desc = texture->GetTextureDescriptor();
+  switch (desc.format) {
+    case impeller::PixelFormat::kB10G10R10XR:  // intentional_fallthrough
+    case impeller::PixelFormat::kR16G16B16A16Float:
+      return ColorSpace::kExtendedSRGB;
+    default:
+      return ColorSpace::kSRGB;
+  }
+}
+
 }  // namespace flutter
